Big performance and scalability increase for protected media
JSON proved as a too simple and insufficiant data structure, since 2 keys (url/id) are needed for instant look-ups (frontend: url & backend: id) to check if media was protected, and the protection details if so.

JSON only allows for 1 key, where we can specify multiple in XML.

After thorough tests of map performance and scalability in go versus slices it was pretty obvious that chosing maps in go and and XML data structure is the way to go.