New Asset/file Manager section
This commit is extremely "beta". It introduces a new section, and a new folder structure for design-related static assets such as css, js and images, which is not content, and should therefore not be stored in the media library.

I'm considering whether the previous Stylesheet & scripts sections should be removed completely, or if they should be used together with or exclusively with database storage of scripts/styles which is to be used to extend the standard features of the CKeditor.