Working on pluging structure
Rewriting my ecommerce & membership WordPress plugin to Collexy.

The point of this is that if such an advanced plugin can be integrated with Collexy, almost any plugin should be possible to hook on the CMS.

While developing the plugin I'll be aware of things I need to change in the core. I'll make the changes as I develop this plugin.